+++ PenisSeeker.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.5.33530.505
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PenisSeeker", "ExamplePlugin\PenisSeeker.csproj", "{38C85AE9-72DC-4B31-9118-3F2CEA660637}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PenisSeeker", "PenisSeeker\PenisSeeker.csproj", "{38C85AE9-72DC-4B31-9118-3F2CEA660637}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
